Modules without type definitions go in their own type definition file.
